sensibull/pyproject.toml
[project]
name = "sensibull-tracker"
version = "0.1.0"
description = "Sensibull Position Tracker"
requires-python = ">=3.9,<3.14"
dependencies = [
    "flask",
    "flask-login",
    "requests",
    "apscheduler",
    "simplejson",
    "anthropic>=0.84.0",
    "python-dotenv>=1.2.1",
    "openai>=2.26.0",
    "psycopg2-binary>=2.9.9",
]
